+++ chessdotcom_ai_coach/routers/components.py
@@ -1,0 +1,4 @@
+from fastapi import APIRouter, Request
+from fastapi.responses import HTMLResponse
+
+router = APIRouter()
