--- chessdotcom_ai_coach/templates/game_list.html
@@ -1,104 +1,0 @@
-{% if not games %}
-<div
-  class="alert alert-warning shadow-lg bg-base-100 border-warning/30 col-span-full"
->
-  <i class="fa-solid fa-circle-info text-2xl"></i>
-  <div>
-    <h3 class="font-bold">No active games</h3>
-    <p class="text-xs">
-      You don't have any ongoing games on Chess.com at the moment.
-    </p>
-  </div>
-</div>
-{% else %} {% for game in games %}
-<div
-  class="card bg-base-100 shadow-xl border border-base-200 hover:border-success/50 transition-all duration-300 group {% if game.is_my_turn %}ring-2 ring-success/50 animate-pulse-slow{% endif %}"
->
-  <div class="card-body p-5">
-    <!-- Game Header -->
-    <div class="flex justify-between items-start mb-4">
-      <div class="flex flex-wrap gap-2">
-        <div
-          class="badge badge-outline badge-sm uppercase font-bold tracking-tighter"
-        >
-          {{ game.time_class }}
-        </div>
-        <div class="badge badge-ghost badge-sm font-mono opacity-70">
-          {{ game.time_control }}
-        </div>
-      </div>
-      {% if game.is_my_turn %}
-      <div
-        class="badge badge-success gap-1 text-[10px] font-black animate-bounce"
-      >
-        <i class="fa-solid fa-play text-[8px]"></i> YOUR TURN
-      </div>
-      {% endif %}
-    </div>
-
-    <!-- Players -->
-    <div class="space-y-3">
-      <!-- White Player -->
-      <div class="flex items-center justify-between">
-        <div class="flex items-center gap-3">
-          <div class="avatar placeholder">
-            <div
-              class="bg-neutral text-neutral-content rounded-full w-8 flex items-center justify-center"
-            >
-              <span class="text-xs">W</span>
-            </div>
-          </div>
-          <div>
-            <div class="font-bold text-sm truncate max-w-[120px]">
-              {{ game.white.username }}
-            </div>
-            <div class="text-[10px] opacity-50 font-mono">
-              {{ game.white.rating }}
-            </div>
-          </div>
-        </div>
-        {% if game.turn == 'white' %}
-        <div class="badge badge-xs badge-success badge-outline">Moving...</div>
-        {% endif %}
-      </div>
-
-      <div class="divider my-0 opacity-10 text-[10px]">VS</div>
-
-      <!-- Black Player -->
-      <div class="flex items-center justify-between">
-        <div class="flex items-center gap-3">
-          <div class="avatar placeholder">
-            <div
-              class="bg-base-300 text-base-content rounded-full w-8 border border-base-content/10 flex items-center justify-center"
-            >
-              <span class="text-xs">B</span>
-            </div>
-          </div>
-          <div>
-            <div class="font-bold text-sm truncate max-w-[120px]">
-              {{ game.black.username }}
-            </div>
-            <div class="text-[10px] opacity-50 font-mono">
-              {{ game.black.rating }}
-            </div>
-          </div>
-        </div>
-        {% if game.turn == 'black' %}
-        <div class="badge badge-xs badge-success badge-outline">Moving...</div>
-        {% endif %}
-      </div>
-    </div>
-
-    <!-- Actions -->
-    <div class="card-actions justify-end mt-6 pt-4 border-t border-base-200">
-      <a
-        href="/detail-board/{{ game.game_id }}"
-        class="btn btn-success btn-sm btn-block group-hover:shadow-lg transition-all"
-      >
-        <i class="fa-solid fa-chess-board"></i>
-        View Board
-      </a>
-    </div>
-  </div>
-</div>
-{% endfor %} {% endif %}
